[Refact] 챔피언리스트 컴포넌트와 탭 컴포넌트 한곳에 모음
